Rebrand SynchronizedMessageIterator to SpecificCountMessageIterator
